refactor(CountdownBar): remove the interface to index file
