--- src/components/feedback/CountdownBar/interfaces/CountdownBar.interface.ts
@@ -1,10 +1,0 @@
-import { AnimationEvent } from "react";
-import { Appearance } from "../types/CountdownBar.Appearance.type";
-
-export interface ICountdownBarProps {
-  size?: string;
-  appearance?: Appearance;
-  duration?: number;
-  isPaused?: boolean;
-  handleCountdown?: (e: AnimationEvent<HTMLDivElement>) => void;
-}
